Restructured and renamed sections.
