+++ src/extended_configparser/configuration/__init__.py
@@ -3,6 +3,6 @@ from __future__ import annotations
 from .configuration import Configuration
 from .entries import ConfigEntry
 from .entries import ConfigEntryCollection
-from .entries import ConfigSection
+from .entries.section import ConfigSection
 
 __all__ = ["Configuration", "ConfigEntryCollection", "ConfigSection", "ConfigEntry"]
