+++ src/extended_configparser/configuration/entries/__init__.py
@@ -2,6 +2,6 @@ from __future__ import annotations
 
 from extended_configparser.configuration.entries.base import ConfigEntry
 from extended_configparser.configuration.entries.base import ConfigEntryCollection
-from extended_configparser.configuration.entries.base import ConfigSection
+from extended_configparser.configuration.entries.section import ConfigSection
 
 __all__ = [ConfigEntry, ConfigEntryCollection, ConfigSection]
