+++ src/extended_configparser/configuration/entries/base.py
@@ -7,6 +7,9 @@ from typing import Any
 
 if TYPE_CHECKING:
     from extended_configparser import ExtendedConfigParser
+    from extended_configparser.configuration.entries.confirmation import (
+        ConfigConfirmationEntry,
+    )
 
 logger = logging.getLogger(__name__)
 
@@ -20,32 +23,6 @@ class ConfigEntryCollection:
     """
 
     pass
-
-
-class ConfigSection:
-    """
-    Helper class to group ConfigEntries together under a section.
-    Create entries by calling `section.ConfigSection("section_name")`.
-    """
-
-    def __init__(self, name: str) -> None:
-        self.name = name
-
-    def ConfigOption(
-        self,
-        option: str,
-        default: str,
-        message: str,
-        **inquirer_kwargs,
-    ) -> ConfigEntry:
-        """Create a ConfigEntry for that section with the given parameters."""
-        return ConfigEntry(
-            section=self.name,
-            option=option,
-            default=default,
-            message=message,
-            **inquirer_kwargs,
-        )
 
 
 class ConfigEntry:
